[TO-1] Correcciones pago multas mercado pago

+++ UI_Internet/src/UI/App.js
@@ -143,7 +143,9 @@ class App extends React.Component {
       if (token == undefined || token == null || token == "undefined" || token == "" || token == window.Config.TOKEN_INVITADO) {
 
         //Borramos el localStorage por si hay algun pago inconcluso o algun dato del redux persistente
-        localStorage.clear();
+        if(!token == window.Config.TOKEN_INVITADO) //Solo en caso que no exista un token
+          localStorage.clear();
+          
         //Logueamos con el usuario Invitado
         this.props.login({
           datos: undefined,
@@ -174,8 +176,6 @@ class App extends React.Component {
               Rules_Usuario.datos(token)
                 .then(datos => {
 
-                  //Borramos el localStorage por si hay algun pago inconcluso o algun dato del redux persistente
-                  localStorage.clear();
                   this.props.login({
                     datos: datos,
                     token: token
